Change CLI renderer to use a different symbol for each bot
The CLI game renderer relies on ANSI colours for distinguising between
players and Spawn points of different bots. However, many consoles
don't correctly support or render ANSI colours, and just print the
whole map in black and white. To help deal with this situation, this
commit introduces a letter for each bot in the game, as well as a
colour. Spawn points are represented by use the upper-case letter
(instead of 'O'), and players are represented by lower-case
(instead of '*').