Make UP-TO-DATE checks for `webpack` task work
Any Gradle task that depends on the viewer build would cause the
`webpack` task to run. This task is slow and expensive, and not
neccessary unless files in the viewer project have changed.
By properly defining the inputs and outputs for this task in the build
file, Gradle's up-to-date checking mechanism will now skip it whenever
possible.