Report reason in viewer if Bot was disqualified
At the end of a game replay, if a Bot was disqualified for some reason,
the reason will be reported in the overlay instaed of
"0 players remaining".

The message of the first `Rejection` used to disqualify a Bot is used
as the 'reason'. These have been tidied up in order to make them a
little more user friendly.

In addition, 'Bot' is no longer stripped from team names in the viewer,
as it has been found that contestants often include this in their team
name intentionally.