Change Owner tag to owner
The tags set up by the business for tracking are mostly capitalised except for owner which has a lowercase 'o'. This PR changes the tag in the various scripts to use the correct casing so it can be tracked properly.

In addition is hooks up some further hard-coded tag values to the values recently added to the config.