Change KeyName to key pair owned by me (parmstrong)
Currently, the existing key pair is owned by Simon Watson from a previous hackathon. This however means only he is able to SSH in and he may not be available during the course of this year's hackathon.

So instead I've created a new key pair through the AWS console and thus have the private key locally stored on my machine. This commit changes the name of the key that is associated with new contestant instances to this new key pair.