HAC-61 Run Gradle with Java 17 (#330)
* Upgrade lombok to 1.18.30

Prior to 1.18.20, lombok would use reflection to gain access to the internal JDK API. Java 16
forbids this so an upgrade is needed.

HAC-61

* Upgrade to groovy 4 for the game project's tests

HAC-61

* Fix PowerMock reflection errors

Since Java 16, it is no longer possible to bypass JPMS using reflection.
As PowerMock heavily relies on reflection, we must explicitly open the
packages it uses by using the JVM argument: `--add-opens`.

PowerMock is no longer maintained so a better longer term solution would
be to remove PowerMock.

HAC-61

* @JsonIgnore GameStateFixture::getAllPositions

`GameStateFixture::getAllPositions` returns a `Stream`. Before Java 16,
this would be serialized to the JSON:
`"allPositions": { "parallel": false }`, which shouldn't be useful.
After Java 16, it causes reflection errors due to JPMS disallowing
reflection access to packages that haven't been opened. Adding
@JsonIgnore is an easy fix.

If we do want to serialize this property then
[Jackon's Java8 modules](https://github.com/FasterXML/jackson-modules-java8)
can be used.

HAC-61

* Run Gradle with Java 17

The source code is still compiled to Java 11 and still needs to be run
using Java 11 but Gradle can now be run using Java 17.

HAC-61

* Add ticket no. to removal of PowerMock comment

HAC-61

* Add Java 11 toolchain to all gradle sub-projects

This will allow JDK 17 to be installed on the machine and still cause
the java code to be compiled to a target version of 11.

HAC-61

* Allow easy testing with the default bots

Added bash script that allows for easy running of the default bots as a
contestant for easy testing.

Each of the default bots now has a constructor that accepts a String to
all the default bots. This allows us to pass in the team name to the
default bots.

HAC-61